Break main.lua into multiple dedicated files
